fix: переименовал классы, одиннадцатый коммит пятнадцатого спринта
