+++ src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -18,7 +18,7 @@ public class ErrorHandler {
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.CONFLICT)
-    public ErrorResponse handleEmailException(final EmailException e) {
+    public ErrorResponse handleEmailException(final EmailAlreadyExistsException e) {
         return new ErrorResponse(
                 String.format("Пользователь с таким email \"%s\" уже существует.", e.getParameter())
         );
