+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.user.service;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exception.EmailException;
+import ru.practicum.shareit.exception.EmailAlreadyExistsException;
 import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.exception.ValidationException;
 import ru.practicum.shareit.user.dto.UserDto;
@@ -72,11 +72,11 @@ public class UserServiceImpl implements UserService {
         }
         if (userDto.getEmail() != null) {
             if (userDto.getEmail().isBlank()) {
-                throw new EmailException("Email не должен быть пустым");
+                throw new ValidationException("Email не должен быть пустым");
             }
             if (!user.getEmail().equals(userDto.getEmail()) && isEmailPresentInRepository(UserMapper
                     .toUser(userDto))) {
-                throw new EmailException(String.format("Email %s уже есть в базе у другого пользователя", userDto.getEmail()));
+                throw new EmailAlreadyExistsException(String.format("Email %s уже есть в базе у другого пользователя", userDto.getEmail()));
             }
             user.setEmail(userDto.getEmail());
         }
