+++ src/test/java/ru/practicum/shareit/item/service/ItemServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking.service;
+package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeEach;
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.status.Status;
+import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.user.model.User;
 
@@ -22,39 +22,41 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 @Transactional
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
 @SpringBootTest
-public class BookingServiceITTest {
-
+public class ItemServiceIntegrationTest {
     private final EntityManager manager;
-    private final BookingService service;
+    private final ItemService service;
 
     private final User owner = new User(null, "John", "john.doe@mail.com");
     private final User booker = new User(null, "Алексей", "alexey.timoshenko@mail.com");
     private final Item item = new Item(null, "Мопед", "Железный конь", true, owner,
             null);
-    private final Booking booking = new Booking(null, LocalDateTime.now().minusMinutes(120),
+    private final Booking bookingLast = new Booking(null, LocalDateTime.now().minusMinutes(120),
             LocalDateTime.now().minusMinutes(60), item, booker, Status.APPROVED);
+    private final Booking bookingNext = new Booking(null, LocalDateTime.now().plusMinutes(60),
+            LocalDateTime.now().plusMinutes(120), item, booker, Status.APPROVED);
 
     @BeforeEach
     void setUp() {
         manager.persist(booker);
         manager.persist(owner);
         manager.persist(item);
-        manager.persist(booking);
+        manager.persist(bookingLast);
+        manager.persist(bookingNext);
     }
 
     @Test
-    void findAllBookingsByUserId() {
-        String state = "ALL";
+    void getItemsByUserId() {
         int from = 0;
         int size = 10;
 
-        List<BookingDto> bookings = service.findAllBookingsByUserId(booker.getId(), state, from, size);
+        List<ItemDto> items = service.getItemsByUserId(owner.getId(), from, size);
 
-        assertNotNull(bookings);
-        assertEquals(1, bookings.size());
-        assertEquals(bookings.get(0).getId(), booking.getId());
-        assertEquals(bookings.get(0).getBooker().getId(), booking.getBooker().getId());
-        assertEquals(bookings.get(0).getItem().getId(), booking.getItem().getId());
-
+        assertNotNull(items);
+        assertEquals(1, items.size());
+        assertNotNull(items.get(0).getLastBooking());
+        assertNotNull(items.get(0).getNextBooking());
+        assertEquals(items.get(0).getId(), item.getId());
+        assertEquals(items.get(0).getName(), item.getName());
+        assertEquals(items.get(0).getDescription(), item.getDescription());
     }
 }
